pyproject.toml
[project]
name = "rag_wixqa"
version = "0.1"
description = "RAG on WixQA dataset"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.26",
    "langchain_openai>=0.3.28",
    "langchain_weaviate>=0.0.5",
    "langgraph_api>=0.2.129",
    "langgraph>=0.6.4",
    "nltk>=3.9.1",
    "evaluate>=0.4.5",
    "datasets>=4.0.0",
    "rouge_score>=0.1.2",
    "sacrebleu>=2.5.1",
    "sqlalchemy>=2.0.41",
    "sqlmodel>=0.0.24",
    "weaviate-client>=4.16.4",
    "pydantic>=2.11.7",
    "arc>=1.0",
    "fastapi",
    "psycopg2",
    "langchain_community>=0.3.27",
    
]

[build-system]
requires = ["setuptools>=78.1.1", "wheel>=0.45.1"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
lint.select = ["E", "F", "I"]
lint.ignore = ["E501"]
exclude = ["migrations", "tests/fixtures", "*.ipynb"] 
target-version = "py311"